Move index.ts to demo folder, and use original index file for library exports
