+++ tsconfig.json
@@ -13,5 +13,4 @@
     "esModuleInterop": true
   },
   "include": ["src/**/*", "declarations"],
-  
 }
